feat(manifest): Group rules by methodology
Refactors the methodology manifest page to group rules under their corresponding methodology.

Previously, the page displayed a flat, ungrouped list of all rules, which made it difficult to see which rules belonged to which methodology at a glance.

This change introduces grouping to provide a more structured and user-friendly view, improving the overall navigation and readability of the manifest.

Signed-off-by: Fred Egbuedike <fredilly@article6.org>